backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic-settings
email-validator
python-multipart
requests 